refactoring some modules
