+++ packages/log/src/index.ts
@@ -1,5 +1,4 @@
 export * from "@spinajs/log-common";
 export * from "./targets/index.js";
 export * from "./log.js";
-export * from "./decorators.js";
 export * from "./bootstrap.js";
