+++ packages/orm/tsconfig.json
@@ -8,11 +8,10 @@
   },
   "include": ["src/**/*"],
   "references": [
-    { "path": "../configuration/tsconfig.json" },
+    { "path": "../configuration-common/tsconfig.json" },
     { "path": "../di/tsconfig.json" },
     { "path": "../exceptions/tsconfig.json" },
     { "path": "../log/tsconfig.json" },
-    { "path": "../reflection/tsconfig.json" }
   ]
 }
