+++ packages/reflection/src/index.ts
@@ -4,34 +4,13 @@ import * as path from 'path';
 import ts from 'typescript';
 import _ from 'lodash';
 import { Configuration } from '@spinajs/configuration';
-import { AsyncService, Class, DI } from '@spinajs/di';
+import { AsyncService, Class, ClassInfo, DI } from '@spinajs/di';
 import { InvalidArgument, Exception } from '@spinajs/exceptions';
 import { Log } from '@spinajs/log';
 
 export * from 'typescript-mix';
 
-/**
- * Class info structure
- */
-export class ClassInfo<T> {
-  /**
-   * Full file path of loaded class
-   */
-  public file: string;
-  /**
-   * Class name
-   */
-  public name: string;
-  /**
-   * Javascript class object
-   */
-  public type: any;
 
-  /**
-   * Resolved instance
-   */
-  public instance?: T;
-}
 
 /**
  * Exception thrown when some error during reflection occurs
